+++ apps/admin-dashboard/app/routes/_dashboard.applications.$id._index.tsx
@@ -12,7 +12,11 @@ import {
 import dayjs from 'dayjs';
 import { type PropsWithChildren, useState } from 'react';
 
-import { getApplication } from '@oyster/core/applications';
+import {
+  acceptApplication,
+  getApplication,
+  rejectApplication,
+} from '@oyster/core/applications';
 import {
   Application as ApplicationType,
   type Gender,
@@ -22,7 +26,6 @@ import {
 } from '@oyster/types';
 import { Button, Text } from '@oyster/ui';
 
-import { acceptApplication, rejectApplication } from '@/admin-dashboard.server';
 import { Application, type EducationLevel } from '@/admin-dashboard.ui';
 import { Route } from '@/shared/constants';
 import {
