+++ apps/admin-dashboard/app/routes/_dashboard.applications.$id.accept.tsx
@@ -10,10 +10,9 @@ import {
   useLoaderData,
 } from '@remix-run/react';
 
-import { getApplication } from '@oyster/core/applications';
+import { acceptApplication, getApplication } from '@oyster/core/applications';
 import { Button, Form, Modal } from '@oyster/ui';
 
-import { acceptApplication } from '@/admin-dashboard.server';
 import { Route } from '@/shared/constants';
 import {
   admin,
